mlu/modules/quantum_ml/quantum_ml.py
class QuantumML:
    def __init__(self):
        pass

    def quantum_algorithm(self):
        """
        Demonstrates a simple quantum algorithm using Qiskit.
        """
        try:
            from qiskit import QuantumCircuit, Aer, execute

            # Create a Quantum Circuit acting on a quantum register of two qubits
            circuit = QuantumCircuit(2)

            # Add a Hadamard gate on qubit 0, putting this qubit in superposition.
            circuit.h(0)
            # Add a CX (CNOT) gate on control qubit 0 and target qubit 1, putting
            # the qubits in a Bell state.
            circuit.cx(0, 1)
            # Map the quantum measurement to the classical bits
            circuit.measure_all()

            # Execute the circuit on the qasm simulator
            simulator = Aer.get_backend('qasm_simulator')
            result = execute(circuit, simulator).result()

            # Return the counts
            counts = result.get_counts(circuit)
            print("Counts:", counts)
        except Exception as e:
            print(f"An error occurred during the quantum algorithm execution: {e}")